Refactored multitenant stuff out further -Remaining setup moved to extension methods.
