+++ src/Finbuckle.MultiTenant.EntityFrameworkCore/Internal/Shared.cs
@@ -27,30 +27,8 @@ namespace Finbuckle.MultiTenant.EntityFrameworkCore
     /// A static class containing static methods shared between
     /// MultiTenantDbContext and MultiTenantIdentityDbContext.
     /// </summary>
-    internal static class Shared
+    public static class Shared
     {
-        /// <summary>
-        /// Adds a "required" constraint to the TenantId string property and
-        /// sets the query filter for the entity on TenantId.
-        /// </summary>
-        /// <remarks>
-        ///     <list type="bullet">
-        ///         <item><description>This method is called from OnModelCreating of the derived DbContext.</description></item>
-        ///         <item><description>If no TenantId string property exists, a shadow property is created.</description></item>
-        ///         <item><description>The query filter has no effect when calling DbSet.Find.</description></item>
-        ///     </list>
-        /// </remarks>
-        /// <param name="modelBuilder"></param>
-        /// <param name="tenantInfo"></param>
-        public static void SetupModel(ModelBuilder modelBuilder)
-        {
-            // Annotate the types marked with the MultiTenant Data Attribute
-            foreach (var t in modelBuilder.Model.GetEntityTypes().Where(t => HasMultiTenantAttribute(t.ClrType)))
-            {
-                modelBuilder.Entity(t.ClrType).IsMultiTenant();
-            }
-        }
-
         internal static LambdaExpression GetQueryFilter(Microsoft.EntityFrameworkCore.Metadata.Builders.EntityTypeBuilder r)
         {
             #if NETSTANDARD2_1
@@ -70,170 +48,6 @@ namespace Finbuckle.MultiTenant.EntityFrameworkCore
         public static bool HasMultiTenantAnnotation(IEntityType t)
         {
             return (bool?)t.FindAnnotation(Constants.MultiTenantAnnotation)?.Value ?? false;
-        }
-
-        /// <summary>
-        /// Checks the TenantId on entities taking into account
-        /// the tenantNotSetMode and the tenantMismatchMode.
-        /// </summary>
-        /// <remarks>
-        ///     <list type="bullet">
-        ///         <item><description>This method is called from SaveChanges or SaveChangesAsync of the derived DbContext.</description></item>
-        ///         <item><description>If any changes are detected in an entitiy with the MultiTenant attribute then the TenantContext must not be null.</description></item>
-        ///     </list>  
-        /// </remarks>
-        /// /// <param name="tenantInfo"></param>
-        /// <param name="changeTracker"></param>
-        /// <param name="tenantNotSetMode"></param>
-        /// <param name="tenantMismatchMode"></param>
-        public static void EnforceTenantId(TenantInfo tenantInfo,
-            ChangeTracker changeTracker,
-            TenantNotSetMode tenantNotSetMode,
-            TenantMismatchMode tenantMismatchMode)
-        {
-            var changedMultiTenantEntities = changeTracker.Entries().
-                Where(e => e.State == EntityState.Added || e.State == EntityState.Modified || e.State == EntityState.Deleted).
-                Where(e => Shared.HasMultiTenantAnnotation(e.Metadata));
-
-            // ensure tenant context is valid
-            if (changedMultiTenantEntities.Any())
-                CheckTenantInfo(tenantInfo);
-
-            // get list of all added entities with MultiTenant annotation
-            var addedMultiTenantEntities = changedMultiTenantEntities.
-                Where(e => e.State == EntityState.Added);
-
-            // handle Tenant Id mismatches for added entities
-            var mismatchedAdded = addedMultiTenantEntities.
-                Where(e => (string)e.Property("TenantId").CurrentValue != null &&
-                (string)e.Property("TenantId").CurrentValue != tenantInfo.Id);
-
-            if (mismatchedAdded.Any())
-            {
-                switch (tenantMismatchMode)
-                {
-                    case TenantMismatchMode.Throw:
-                        throw new MultiTenantException($"{mismatchedAdded.Count()} added entities with Tenant Id mismatch."); ;
-
-                    case TenantMismatchMode.Ignore:
-                        // no action needed
-                        break;
-
-                    case TenantMismatchMode.Overwrite:
-                        foreach (var e in mismatchedAdded)
-                        {
-                            e.Property("TenantId").CurrentValue = tenantInfo.Id;
-                        }
-                        break;
-                }
-            }
-
-            // for added entities TenantNotSetMode is always Overwrite
-            var notSetAdded = addedMultiTenantEntities.
-                Where(e => (string)e.Property("TenantId").CurrentValue == null);
-
-            foreach (var e in notSetAdded)
-            {
-                e.Property("TenantId").CurrentValue = tenantInfo.Id;
-            }
-
-            // get list of all modified entities with MultiTenant annotation
-            var modifiedMultiTenantEntities = changedMultiTenantEntities.
-                Where(e => e.State == EntityState.Modified);
-
-            // handle Tenant Id mismatches for modified entities
-            var mismatchedModified = modifiedMultiTenantEntities.
-                Where(e => (string)e.Property("TenantId").CurrentValue != null &&
-                (string)e.Property("TenantId").CurrentValue != tenantInfo.Id);
-
-            if (mismatchedModified.Any())
-            {
-                switch (tenantMismatchMode)
-                {
-                    case TenantMismatchMode.Throw:
-                        throw new MultiTenantException($"{mismatchedModified.Count()} modified entities with Tenant Id mismatch."); ;
-
-                    case TenantMismatchMode.Ignore:
-                        // no action needed
-                        break;
-
-                    case TenantMismatchMode.Overwrite:
-                        foreach (var e in mismatchedModified)
-                        {
-                            e.Property("TenantId").CurrentValue = tenantInfo.Id;
-                        }
-                        break;
-                }
-            }
-
-            // handle Tenant Id not set for modified entities
-            var notSetModified = modifiedMultiTenantEntities.
-                Where(e => (string)e.Property("TenantId").CurrentValue == null);
-
-            if (notSetModified.Any())
-            {
-                switch (tenantNotSetMode)
-                {
-                    case TenantNotSetMode.Throw:
-                        throw new MultiTenantException($"{notSetModified.Count()} modified entities with Tenant Id not set."); ;
-
-                    case TenantNotSetMode.Overwrite:
-                        foreach (var e in notSetModified)
-                        {
-                            e.Property("TenantId").CurrentValue = tenantInfo.Id;
-                        }
-                        break;
-                }
-            }
-
-            // get list of all deleted  entities with MultiTenant annotation
-            var deletedMultiTenantEntities = changedMultiTenantEntities.
-                Where(e => e.State == EntityState.Deleted);
-
-            // handle Tenant Id mismatches for deleted entities
-            var mismatchedDeleted = deletedMultiTenantEntities.
-                Where(e => (string)e.Property("TenantId").CurrentValue != null &&
-                (string)e.Property("TenantId").CurrentValue != tenantInfo.Id);
-
-            if (mismatchedDeleted.Any())
-            {
-                switch (tenantMismatchMode)
-                {
-                    case TenantMismatchMode.Throw:
-                        throw new MultiTenantException($"{mismatchedDeleted.Count()} deleted entities with Tenant Id mismatch."); ;
-
-                    case TenantMismatchMode.Ignore:
-                        // no action needed
-                        break;
-
-                    case TenantMismatchMode.Overwrite:
-                        // no action needed
-                        break;
-                }
-            }
-
-            // handle Tenant Id not set for deleted entities
-            var notSetDeleted = deletedMultiTenantEntities.
-                Where(e => (string)e.Property("TenantId").CurrentValue == null);
-
-            if (notSetDeleted.Any())
-            {
-                switch (tenantNotSetMode)
-                {
-                    case TenantNotSetMode.Throw:
-                        throw new MultiTenantException($"{notSetDeleted.Count()} deleted entities with Tenant Id not set."); ;
-
-                    case TenantNotSetMode.Overwrite:
-                        // no action needed
-                        break;
-                }
-            }
-        }
-
-        private static void CheckTenantInfo(TenantInfo tenantInfo)
-        {
-            if (tenantInfo == null)
-                throw new MultiTenantException("MultiTenant Entity cannot be changed if TenantInfo is null.");
         }
     }
 }
